refactor to avoid circular dependencies
