+++ main.py
@@ -1,15 +1,19 @@
 from pathlib import Path
 
+from tenacity import retry_if_exception_type
+
 from globals.bulk_processor import bulk_process_videos
 from globals.io.dataclasses import read_json_to_dataclass
 from globals.io.paths import PathHelper, get_videos_folder_path
-from globals.physics_processor import PhysicsProcessor, UserInput
+from globals.physics_processor import PhysicsProcessor
+from globals.user_input import UserInput
 from globals.video_display import display
 
-# if __name__ == '__main__':
-#     bulk_process_videos()
-
 if __name__ == '__main__':
+    bulk_process_videos()
+
+    exit(0)
+# if __name__ == '__main__':
     path = get_videos_folder_path()
     for video in path.iterdir():
         path_helper = PathHelper(Path(video))
